pyproject.toml
[project]
name = "multi_agent_crew"
version = "0.1.0"
description = "Multi-agent crew using crewAI for weather and SEO analysis"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.6.1",
    "requests>=2.31.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.100.0",
    "openai>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic-settings>=2.0.0",
    "prometheus-client>=0.17.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1"
]

[project.scripts]
multi_agent_crew = "multi_agent_crew.main:run"
run_crew = "multi_agent_crew.main:run"
train = "multi_agent_crew.main:train"
replay = "multi_agent_crew.main:replay"
test = "multi_agent_crew.main:test"
run_with_trigger = "multi_agent_crew.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
